// DetailListGrid/types/DateFilterOperator.ts
export enum DateFilterOperator {
    Date = 'date',
    On_or_after = 'onafter',
    On_or_before = 'onbefore',
    Today = 'today',
    This_week = 'thisweek',
    This_month = 'thismonth'
}
//   const options = [
//     { key: 'date', text: 'Datum' },
//     { key: 'onafter', text: 'På eller efter' },
//     { key: 'today', text: 'Idag' },
//   ];